+++ lib/teleprompter.dart
@@ -1,4 +1,6 @@
 library teleprompter;
 
+export 'package:teleprompter/src/data/capture_mode.dart';
 export 'package:teleprompter/src/teleprompter_widget.dart';
+export 'package:teleprompter/src/ui/camera/recording_marker.dart';
 export 'package:teleprompter/src/ui/timer/stopwatch_widget.dart';